@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,13 +26,13 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  retrofit: ^1.3.4+1
+  retrofit: ^2.0.0-beta1
   logger: any
-  json_annotation: ^3.1.0
-  json_serializable: ^3.5.1
+  json_annotation: ^4.0.1
+  json_serializable: ^4.1.0
   intl: ^0.17.0
-  flutter_riverpod: ^0.12.1
-  hooks_riverpod: ^0.12.1
+  flutter_riverpod: ^0.14.0+1
+  hooks_riverpod: ^0.14.0+1
   hive: ^2.0.0
   hive_flutter: ^1.0.0
   firebase_core: ^1.0.2
@@ -53,7 +53,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  retrofit_generator: ^1.4.1+3
+  retrofit_generator: ^2.0.0-beta3
   hive_generator: ^1.0.0
   build_runner: any
 
