@@ -10,48 +10,45 @@ import 'package:timezone/timezone.dart' as tz;
 import 'package:weather_app/models/recieved_notification.dart';
 
 final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-FlutterLocalNotificationsPlugin();
+    FlutterLocalNotificationsPlugin();
 
-/// Streams are created so that app can respond to notification-related events
-/// since the plugin is initialised in the `main` function
 final BehaviorSubject<ReceivedNotification> didReceiveLocalNotificationSubject =
-BehaviorSubject<ReceivedNotification>();
+    BehaviorSubject<ReceivedNotification>();
 
 final BehaviorSubject<String> selectNotificationSubject =
-BehaviorSubject<String>();
+    BehaviorSubject<String>();
 
-const MethodChannel platform =
-MethodChannel('com.example/weather_app');
+const MethodChannel platform = MethodChannel('com.example/weather_app');
 
-String selectedNotificationPayload;
-
+String? selectedNotificationPayload;
 
 class NotificationSettingsModel {
+  final DateTime? selectedDateTime;
+
   const NotificationSettingsModel(this.selectedDateTime);
-  final DateTime selectedDateTime;
 
   Future<void> scheduleDailyNotification() async {
     await flutterLocalNotificationsPlugin.zonedSchedule(
-        0,
-        'Weather',
-        'Check the weather forecast before going outside',
-        _nextInstanceOfSelectedTime(selectedDateTime),
-        const NotificationDetails(
-          android: AndroidNotificationDetails(
-              'daily notification channel id',
-              'daily notification channel name',
-              'daily notification description'),
-        ),
-        androidAllowWhileIdle: true,
-        uiLocalNotificationDateInterpretation:
-        UILocalNotificationDateInterpretation.absoluteTime,
-        matchDateTimeComponents: DateTimeComponents.time);
+      0,
+      'Weather',
+      'Check the weather forecast before going outside',
+      _nextInstanceOfSelectedTime(selectedDateTime!),
+      const NotificationDetails(
+        android: AndroidNotificationDetails(
+            'daily notification channel id',
+            'daily notification channel name',
+            'daily notification description'),
+      ),
+      androidAllowWhileIdle: true,
+      uiLocalNotificationDateInterpretation:
+          UILocalNotificationDateInterpretation.absoluteTime,
+      matchDateTimeComponents: DateTimeComponents.time,
+    );
   }
 
   tz.TZDateTime _nextInstanceOfSelectedTime(DateTime dateTime) {
-    tz.TZDateTime scheduledDate =
-    tz.TZDateTime(tz.local, dateTime.year, dateTime.month, dateTime.day,
-        dateTime.hour, dateTime.minute);
+    tz.TZDateTime scheduledDate = tz.TZDateTime(tz.local, dateTime.year,
+        dateTime.month, dateTime.day, dateTime.hour, dateTime.minute,);
     if (scheduledDate.isBefore(dateTime)) {
       scheduledDate = scheduledDate.add(const Duration(days: 1));
     }
@@ -63,41 +60,30 @@ class NotificationSettingsModel {
   }
 }
 
-Future<void> setupNotifications() async{
+Future<void> setupNotifications() async {
   await _configureLocalTimeZone();
 
-  // TODO: Find out where it was used
-  // final NotificationAppLaunchDetails notificationAppLaunchDetails =
-  // await flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails();
-  // String initialRoute = HomePage.routeName;
-  // if (notificationAppLaunchDetails?.didNotificationLaunchApp ?? false) {
-  //   selectedNotificationPayload = notificationAppLaunchDetails!.payload;
-  //   initialRoute = SecondPage.routeName;
-  // }
-
   const AndroidInitializationSettings initializationSettingsAndroid =
-  AndroidInitializationSettings('app_icon');
+      AndroidInitializationSettings('app_icon');
 
-  /// Note: permissions aren't requested here just to demonstrate that can be
-  /// done later
   final InitializationSettings initializationSettings = InitializationSettings(
     android: initializationSettingsAndroid,
   );
   await flutterLocalNotificationsPlugin.initialize(initializationSettings,
-      onSelectNotification: (String payload) async {
-        if (payload != null) {
-          debugPrint('notification payload: $payload');
-        }
-        selectedNotificationPayload = payload;
-        selectNotificationSubject.add(payload);
-      });
+      onSelectNotification: (String? payload) async {
+    if (payload != null) {
+      debugPrint('notification payload: $payload');
+    }
+    selectedNotificationPayload = payload!;
+    selectNotificationSubject.add(payload);
+  });
 }
 
 Future<void> _configureLocalTimeZone() async {
   tz.initializeTimeZones();
   final String timeZoneName =
-  await platform.invokeMethod<String>('getTimeZoneName');
-  tz.setLocalLocation(tz.getLocation(timeZoneName));       //changed '!' after timeZoneName
+      (await platform.invokeMethod<String>('getTimeZoneName'))!;
+  tz.setLocalLocation(tz.getLocation(timeZoneName));
 }
 
 void dispose() {
