@@ -1,27 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:weather_app/models/hourly_weather.dart';
 import 'package:weather_app/notifiers/hourly_weather_notifier.dart';
 import 'package:weather_app/ui/views/full_hour_description.dart';
 import 'package:weather_app/ui/views/load_button.dart';
 import 'package:weather_app/localization/app_localizations.dart';
-
 
 final hourlyWeatherProvider = StateNotifierProvider(
   (ref) => HourlyWeatherNotifier()..loadDataFromHive(),
 );
 
 class HourlyWeatherPage extends HookWidget {
-  const HourlyWeatherPage({Key key}) : super(key: key);
-
-  Widget _buildBody(HourlyWeather hourlyWeather) {
-    return FullHourDescription(hourlyWeather: hourlyWeather,);
-  }
+  const HourlyWeatherPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final hourlyWeather = useProvider(hourlyWeatherProvider.state);
+    final hourlyWeather = useProvider(hourlyWeatherProvider);
 
     return Scaffold(
       appBar: AppBar(
@@ -34,13 +28,15 @@ class HourlyWeatherPage extends HookWidget {
         ),
       ),
       floatingActionButton: LoadButton(
-        onPressed: context.read(hourlyWeatherProvider).loadDataFromApi,
+        onPressed: context.read(hourlyWeatherProvider.notifier).loadDataFromApi,
       ),
       body: hourlyWeather == null
           ? Center(
               child: CircularProgressIndicator(),
             )
-          : _buildBody(hourlyWeather),
+          : FullHourDescription(
+              hourlyWeather: hourlyWeather,
+            ),
     );
   }
 }
