@@ -5,21 +5,22 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:weather_app/localization/app_language.dart';
 import 'package:weather_app/localization/app_localizations.dart';
-import 'package:weather_app/notifiers/notification_settings_notifier.dart';
 import 'package:weather_app/ui/views/padded_elevated_button.dart';
+import 'package:weather_app/notifiers/notification_settings_notifier.dart';
+import 'package:weather_app/models/notification_settings_model.dart';
 import 'package:weather_app/utils/date_utils.dart';
 
-final notificationSettingsProvider =
-    StateNotifierProvider((ref) => NotificationSettingsNotifier());
+final notificationSettingsProvider = StateNotifierProvider(
+  (ref) => NotificationSettingsNotifier(),
+);
 
 class NotificationSettingsPage extends HookWidget {
-  const NotificationSettingsPage({Key key}) : super(key: key);
+  const NotificationSettingsPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final notificationSettingsModel = useProvider(
-      notificationSettingsProvider.state,
-    );
+    final notificationSettingsModel =
+        useProvider(notificationSettingsProvider) as NotificationSettingsModel;
 
     var appLanguage = getAppLanguage(context);
     return Scaffold(
@@ -28,13 +29,15 @@ class NotificationSettingsPage extends HookWidget {
           AppLocalizations.of(context).translate('notification_settings'),
         ),
         leading: IconButton(
-          icon: Icon(Icons.arrow_back),
+          icon: Icon(
+            Icons.arrow_back,
+          ),
           onPressed: () => Navigator.of(context).pop(),
         ),
       ),
       body: SingleChildScrollView(
         child: Padding(
-          padding: const EdgeInsets.all(8),
+          padding: EdgeInsets.all(8),
           child: Center(
             child: Column(
               children: <Widget>[
@@ -48,7 +51,7 @@ class NotificationSettingsPage extends HookWidget {
                       notificationSettingsModel.selectedDateTime == null
                           ? CircularProgressIndicator()
                           : Text(
-                              notificationSettingsModel.selectedDateTime
+                              notificationSettingsModel.selectedDateTime!
                                   .getFormattedFullDateTimeString(),
                             ),
                       SizedBox(
@@ -56,7 +59,7 @@ class NotificationSettingsPage extends HookWidget {
                       ),
                       ElevatedButton(
                         onPressed: () => context
-                            .read(notificationSettingsProvider)
+                            .read(notificationSettingsProvider.notifier)
                             .selectDate(context),
                         child: Text(
                           AppLocalizations.of(context)
@@ -87,19 +90,19 @@ class NotificationSettingsPage extends HookWidget {
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: <Widget>[
                     ElevatedButton(
-                      onPressed: () {
+                      onPressed: () async {
                         appLanguage.changeLanguage(Locale("en"));
                       },
                       child: Text('English'),
                     ),
                     ElevatedButton(
-                      onPressed: () {
+                      onPressed: () async {
                         appLanguage.changeLanguage(Locale("ru"));
                       },
                       child: Text('Русский'),
-                    )
+                    ),
                   ],
-                )
+                ),
               ],
             ),
           ),
