@@ -1,4 +1,4 @@
-class Constants{
+class Constants {
   static const String notificationSettings = 'notification_settings';
   static const String hourlyWeather = 'hourly_weather';
   static const List<String> choices = <String>[
