@@ -20,13 +20,13 @@ class DayAdapter extends TypeAdapter<Day> {
       dt: fields[0] as int,
       sunrise: fields[1] as int,
       sunset: fields[2] as int,
+      temp: fields[3] as Temperature,
       pressure: fields[4] as int,
       humidity: fields[5] as int,
-      wind_speed: fields[6] as double,
+      windSpeed: fields[6] as double,
+      weather: (fields[7] as List).cast<Weather>(),
       clouds: fields[8] as int,
-    )
-      ..temp = fields[3] as Temperature
-      ..weather = (fields[7] as List)?.cast<Weather>();
+    );
   }
 
   @override
@@ -46,7 +46,7 @@ class DayAdapter extends TypeAdapter<Day> {
       ..writeByte(5)
       ..write(obj.humidity)
       ..writeByte(6)
-      ..write(obj.wind_speed)
+      ..write(obj.windSpeed)
       ..writeByte(7)
       ..write(obj.weather)
       ..writeByte(8)
@@ -73,28 +73,25 @@ Day _$DayFromJson(Map<String, dynamic> json) {
     dt: json['dt'] as int,
     sunrise: json['sunrise'] as int,
     sunset: json['sunset'] as int,
+    temp: Temperature.fromJson(json['temp'] as Map<String, dynamic>),
     pressure: json['pressure'] as int,
     humidity: json['humidity'] as int,
-    wind_speed: (json['wind_speed'] as num)?.toDouble(),
+    windSpeed: (json['wind_speed'] as num).toDouble(),
+    weather: (json['weather'] as List<dynamic>)
+        .map((e) => Weather.fromJson(e as Map<String, dynamic>))
+        .toList(),
     clouds: json['clouds'] as int,
-  )
-    ..temp = json['temp'] == null
-        ? null
-        : Temperature.fromJson(json['temp'] as Map<String, dynamic>)
-    ..weather = (json['weather'] as List)
-        ?.map((e) =>
-            e == null ? null : Weather.fromJson(e as Map<String, dynamic>))
-        ?.toList();
+  );
 }
 
 Map<String, dynamic> _$DayToJson(Day instance) => <String, dynamic>{
       'dt': instance.dt,
       'sunrise': instance.sunrise,
       'sunset': instance.sunset,
-      'temp': instance.temp?.toJson(),
+      'temp': instance.temp.toJson(),
       'pressure': instance.pressure,
       'humidity': instance.humidity,
-      'wind_speed': instance.wind_speed,
-      'weather': instance.weather?.map((e) => e?.toJson())?.toList(),
+      'wind_speed': instance.windSpeed,
+      'weather': instance.weather.map((e) => e.toJson()).toList(),
       'clouds': instance.clouds,
     };
