@@ -1,5 +1,5 @@
-import 'package:get_it/get_it.dart';
 import 'package:dio/dio.dart';
+import 'package:get_it/get_it.dart';
 import 'package:weather_app/api/rest_client.dart';
 
 final getIt = GetIt.instance;
