@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 
-class LoadButton extends StatelessWidget{
-  const LoadButton({Key key, this.onPressed}) : super(key: key);
+class LoadButton extends StatelessWidget {
   final GestureTapCallback onPressed;
 
+  const LoadButton({Key? key, required this.onPressed}) : super(key: key);
+
   @override
-  Widget build(BuildContext context){
+  Widget build(BuildContext context) {
     return FloatingActionButton(
       onPressed: onPressed,
       child: Icon(Icons.cloud_download),
